requirements.txt
streamlit==1.40.2
langchain==0.3.9
langchain_community==0.3.9
langchain-google-genai==2.0.6
google-search-results==2.4.2
google-cloud-bigquery====3.27.0